export violation component to be used in WGE
