+++ ui/App.tsx
@@ -44,7 +44,7 @@ import OCIRepositoryPage from "./pages/v2/OCIRepositoryPage";
 import ProviderPage from "./pages/v2/ProviderPage";
 import Sources from "./pages/v2/Sources";
 import UserInfo from "./pages/v2/UserInfo";
-import PolicyViolationDetails from "./components/Policies/PolicyViolations/PolicyViolationDetails";
+import PolicyViolationPage from "./components/Policies/PolicyViolations/PolicyViolationPage";
 
 const queryClient = new QueryClient();
 
@@ -120,7 +120,7 @@ const App = () => {
           />
           <Route
             path={V2Routes.PolicyViolationDetails}
-            component={withSearchParams(PolicyViolationDetails)}
+            component={withSearchParams(PolicyViolationPage)}
           />
           <Route path={V2Routes.UserInfo} component={UserInfo} />
