+++ ui/components/Policies/PolicyViolations/PolicyViolationDetails.tsx
@@ -2,19 +2,14 @@ import React from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 
-import styled from "styled-components";
-import { useGetPolicyValidationDetails } from "../../../hooks/policyViolations";
 import { PolicyValidation } from "../../../lib/api/core/core.pb";
-import { Kind } from "../../../lib/api/core/types.pb";
-import { formatURL } from "../../../lib/nav";
-import { V2Routes } from "../../../lib/types";
 import Flex from "../../Flex";
 import Link from "../../Link";
-import Page from "../../Page";
 import Text from "../../Text";
 import Timestamp from "../../Timestamp";
 
 import { AppContext } from "../../../contexts/AppContext";
+import { useFeatureFlags } from "../../../hooks/featureflags";
 import { FluxObject } from "../../../lib/objects";
 import HeaderRows, { Header } from "../Utils/HeaderRows";
 import Parameters from "../Utils/Parameters";
@@ -36,11 +31,12 @@ interface ViolationDetailsProps {
   entityUrl: string;
   entityObject: FluxObject;
 }
-const ViolationDetails = ({
+export const ViolationDetails = ({
   violation,
   entityObject,
   entityUrl,
 }: ViolationDetailsProps) => {
+  const { isFlagEnabled } = useFeatureFlags();
   const { setDetailModal } = React.useContext(AppContext);
   const {
     severity,
@@ -59,6 +55,11 @@ const ViolationDetails = ({
     {
       rowkey: "Policy Name",
       value: name,
+    },
+    {
+      rowkey: "Cluster",
+      value: "clusterName",
+      visible: isFlagEnabled("WEAVE_GITOPS_FEATURE_CLUSTER"),
     },
     {
       rowkey: "Application",
@@ -123,88 +124,3 @@ const ViolationDetails = ({
     </Flex>
   );
 };
-
-interface Props {
-  id: string;
-  name: string;
-  clusterName?: string;
-  className?: string;
-  kind?: string;
-}
-
-const PolicyViolationDetails = ({ id, name, className, kind }: Props) => {
-  const { data, error, isLoading } = useGetPolicyValidationDetails({
-    violationId: id,
-  });
-
-  const violation = data?.violation;
-  const entityObject = new FluxObject({
-    payload: violation?.violatingEntity,
-  });
-
-  const entityUrl = formatURL(
-    Kind[kind] === Kind.Kustomization
-      ? `${V2Routes.Kustomization}/violations`
-      : `${V2Routes.HelmRelease}/violations`,
-    {
-      name: violation?.entity,
-      namespace: violation?.namespace,
-      clusterName: violation?.clusterName,
-    }
-  );
-  return (
-    <Page
-      error={error}
-      loading={isLoading}
-      className={className}
-      path={[
-        { label: "Applications", url: V2Routes.Automations },
-        { label: violation?.entity, url: entityUrl },
-        { label: name || "" },
-      ]}
-    >
-      {data && (
-        <ViolationDetails
-          violation={data.violation}
-          entityUrl={entityUrl}
-          entityObject={entityObject}
-        />
-      )}
-    </Page>
-  );
-};
-
-export default styled(PolicyViolationDetails)`
-  .editor {
-    & a {
-      color: ${(props) => props.theme.colors.primary};
-    }
-    ,
-    & > *:first-child {
-      margin-top: 0;
-    }
-    ,
-    & > *:last-child {
-      margin-bottom: 0;
-    }
-
-    width: calc(100% - 24px);
-    padding: 12px;
-    overflow: scroll;
-    background: ${(props) => props.theme.colors.neutralGray};
-    max-height: 300px;
-  }
-  .code {
-    pre {
-      max-height: 300px;
-      overflow: auto;
-    }
-    code > span {
-      flex-wrap: wrap;
-    }
-  }
-  ul.occurrences {
-    padding-left: ${(props) => props.theme.spacing.base};
-    margin: 0;
-  }
-`;
